# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PetPlants API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type EmoticonStyle = typeof EmoticonStyle[keyof typeof EmoticonStyle];


export const EmoticonStyle = {
  leafy: 'leafy',
  succulent: 'succulent',
  flower: 'flower',
  herb: 'herb',
} as const;

export type PlantState = typeof PlantState[keyof typeof PlantState];


export const PlantState = {
  happy: 'happy',
  thirsty: 'thirsty',
  resting: 'resting',
  postponed: 'postponed',
} as const;

export interface Plant {
  id: number;
  name: string;
  /** @nullable */
  species?: string | null;
  frequencyDays: number;
  /** @nullable */
  waterAmount?: string | null;
  /** @nullable */
  notes?: string | null;
  emoticonStyle: EmoticonStyle;
  createdAt: string;
  /** @nullable */
  nextWaterDate?: string | null;
  /** @nullable */
  lastWateredDate?: string | null;
  state: PlantState;
  dueToday: boolean;
  totalXp: number;
}

export interface PlantInput {
  /** @minLength 1 */
  name: string;
  species?: string;
  /** @minimum 1 */
  frequencyDays: number;
  waterAmount?: string;
  notes?: string;
  emoticonStyle?: EmoticonStyle;
}

export interface PlantUpdate {
  /** @minLength 1 */
  name?: string;
  species?: string;
  /** @minimum 1 */
  frequencyDays?: number;
  waterAmount?: string;
  notes?: string;
  emoticonStyle?: EmoticonStyle;
}

export interface PlayerStats {
  totalXp: number;
  currentLevel: number;
  levelTitle: string;
  currentStreak: number;
  longestStreak: number;
  /** @nullable */
  xpToNextLevel?: number | null;
  xpForCurrentLevel: number;
}

export interface PlantCreatedResult {
  plant: Plant;
  xpAwarded: number;
  stats: PlayerStats;
}

export type WateringLogInputStatus = typeof WateringLogInputStatus[keyof typeof WateringLogInputStatus];


export const WateringLogInputStatus = {
  watered: 'watered',
  postponed: 'postponed',
} as const;

export interface WateringLogInput {
  plantId: number;
  status: WateringLogInputStatus;
  /** ISO date string YYYY-MM-DD */
  logDate?: string;
}

export interface WateringLogResult {
  xpAwarded: number;
  stats: PlayerStats;
  plant: Plant;
  streakBonus: number;
  leveledUp: boolean;
}

/**
 * @nullable
 */
export type HistoryEntryStatus = typeof HistoryEntryStatus[keyof typeof HistoryEntryStatus] | null;


export const HistoryEntryStatus = {
  watered: 'watered',
  postponed: 'postponed',
  missed: 'missed',
} as const;

export interface HistoryEntry {
  date: string;
  /** @nullable */
  status?: HistoryEntryStatus;
  /** @nullable */
  xpAwarded?: number | null;
  /** @nullable */
  notes?: string | null;
}

export interface TodaySummary {
  plantsDueCount: number;
  plantsWateredToday: number;
  totalPlants: number;
}

export interface Dashboard {
  plants: Plant[];
  summary: TodaySummary;
  stats: PlayerStats;
}