# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PetPlants API
 * OpenAPI spec version: 0.1.0
 */
import type { Plant } from './plant';
import type { PlayerStats } from './playerStats';
import type { TodaySummary } from './todaySummary';

export interface Dashboard {
  plants: Plant[];
  summary: TodaySummary;
  stats: PlayerStats;
}