# lib/api-zod/src/generated/types/historyEntry.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PetPlants API
 * OpenAPI spec version: 0.1.0
 */
import type { HistoryEntryStatus } from './historyEntryStatus';

export interface HistoryEntry {
  date: string;
  /** @nullable */
  status?: HistoryEntryStatus;
  /** @nullable */
  xpAwarded?: number | null;
  /** @nullable */
  notes?: string | null;
}