# lib/api-zod/src/generated/types/plant.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PetPlants API
 * OpenAPI spec version: 0.1.0
 */
import type { EmoticonStyle } from './emoticonStyle';
import type { PlantState } from './plantState';

export interface Plant {
  id: number;
  name: string;
  /** @nullable */
  species?: string | null;
  frequencyDays: number;
  /** @nullable */
  waterAmount?: string | null;
  /** @nullable */
  notes?: string | null;
  emoticonStyle: EmoticonStyle;
  createdAt: string;
  /** @nullable */
  nextWaterDate?: string | null;
  /** @nullable */
  lastWateredDate?: string | null;
  state: PlantState;
  dueToday: boolean;
  totalXp: number;
}