# lib/api-zod/src/generated/types/plantState.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PetPlants API
 * OpenAPI spec version: 0.1.0
 */

export type PlantState = typeof PlantState[keyof typeof PlantState];


export const PlantState = {
  happy: 'happy',
  thirsty: 'thirsty',
  resting: 'resting',
  postponed: 'postponed',
} as const;