# lib/api-zod/src/generated/types/playerStats.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PetPlants API
 * OpenAPI spec version: 0.1.0
 */

export interface PlayerStats {
  totalXp: number;
  currentLevel: number;
  levelTitle: string;
  currentStreak: number;
  longestStreak: number;
  /** @nullable */
  xpToNextLevel?: number | null;
  xpForCurrentLevel: number;
}