# lib/api-zod/src/generated/types/todaySummary.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PetPlants API
 * OpenAPI spec version: 0.1.0
 */

export interface TodaySummary {
  plantsDueCount: number;
  plantsWateredToday: number;
  totalPlants: number;
}