# lib/api-zod/src/generated/types/wateringLogInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PetPlants API
 * OpenAPI spec version: 0.1.0
 */
import type { WateringLogInputStatus } from './wateringLogInputStatus';

export interface WateringLogInput {
  plantId: number;
  status: WateringLogInputStatus;
  /** ISO date string YYYY-MM-DD */
  logDate?: string;
}