# lib/api-zod/src/generated/types/wateringLogResult.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * PetPlants API
 * OpenAPI spec version: 0.1.0
 */
import type { Plant } from './plant';
import type { PlayerStats } from './playerStats';

export interface WateringLogResult {
  xpAwarded: number;
  stats: PlayerStats;
  plant: Plant;
  streakBonus: number;
  leveledUp: boolean;
}